+++ libs/commons/src/main/java/com/recom/commons/map/rasterizer/SlopeAndAspectMapRasterizer.java
@@ -28,7 +28,7 @@ public class SlopeAndAspectMapRasterizer implements MapLayerRasterizer {
 
     @NonNull
     public SlopeAndAspectMap rasterizeSlopeAndAspectMap(@NonNull final DEMDescriptor demDescriptor) {
-        final D8AlgorithmForSlopeAndAspectMap algorithmForSlopeAndAspect = new D8AlgorithmForSlopeAndAspectMap(demDescriptor.getStepSize());
+        final D8AlgorithmForSlopeAndAspectMap algorithmForSlopeAndAspect = new D8AlgorithmForSlopeAndAspectMap(demDescriptor.getStepSize().doubleValue());
         final SlopeAndAspect[][] slopeAndAspects = algorithmForSlopeAndAspect.generateSlopeAndAspectMap(demDescriptor.getDem());
 
         return SlopeAndAspectMap.builder()
